Refactor repository models to centralize database operation handling: replace direct imports of `safeDbOperation` with a helper function in each repository file. Update link, tag, and user update methods to use Prisma input types for improved type safety and maintainability.
