+++ src/lib/models/link-repository.ts
@@ -1,9 +1,25 @@
-import { prisma, safeDbOperation } from "./db";
+import { prisma } from "@/lib/prisma";
 import { Link } from "./index";
 import { createLinkSchema } from "./schema";
 import { z } from "zod";
+import { Prisma } from '@/generated/prisma';
 
 export type CreateLinkInput = z.infer<typeof createLinkSchema>;
+
+/**
+ * Helper function to safely execute database operations
+ */
+async function safeDbOperation<T>(
+  operation: () => Promise<T>
+): Promise<{ data: T | null; error: Error | null }> {
+  try {
+    const data = await operation();
+    return { data, error: null };
+  } catch (error) {
+    console.error("Database error:", error);
+    return { data: null, error: error instanceof Error ? error : new Error("An unknown database error occurred") };
+  }
+}
 
 /**
  * Repository for Link model operations
@@ -96,7 +112,7 @@ export const LinkRepository = {
   /**
    * Update a link
    */
-  async update(id: string, data: Partial<Link>) {
+  async update(id: string, data: Prisma.LinkUpdateInput) {
     return safeDbOperation(() => 
       prisma.link.update({
         where: { id },
