+++ src/lib/models/user-repository.ts
@@ -1,9 +1,24 @@
-import { prisma, safeDbOperation } from "./db";
+import { prisma } from "@/lib/prisma";
 import { User } from "./index";
 import { createUserSchema } from "./schema";
 import { z } from "zod";
 
 export type CreateUserInput = z.infer<typeof createUserSchema>;
+
+/**
+ * Helper function to safely execute database operations
+ */
+async function safeDbOperation<T>(
+  operation: () => Promise<T>
+): Promise<{ data: T | null; error: Error | null }> {
+  try {
+    const data = await operation();
+    return { data, error: null };
+  } catch (error) {
+    console.error("Database error:", error);
+    return { data: null, error: error instanceof Error ? error : new Error("An unknown database error occurred") };
+  }
+}
 
 /**
  * Repository for User model operations
@@ -45,7 +60,12 @@ export const UserRepository = {
   /**
    * Update a user
    */
-  async update(id: string, data: Partial<User>) {
+  async update(id: string, data: {
+    email?: string;
+    name?: string;
+    image?: string | null;
+    updatedAt?: Date;
+  }) {
     return safeDbOperation(() => 
       prisma.user.update({
         where: { id },
